fix: remove waitForElmEx to wait_elm.ts
